refactor(client): ♻️ split firebase into multiple files
