+++ client/src/firebase/app.ts
@@ -1,0 +1,12 @@
+import { initializeApp } from "firebase/app";
+
+const firebaseConfig = {
+  apiKey: "AIzaSyDlCorjwKWeIT27bfClJljDGLVv__9PaRw",
+  authDomain: "hedge-site.firebaseapp.com",
+  projectId: "hedge-site",
+  storageBucket: "hedge-site.appspot.com",
+  messagingSenderId: "662700376854",
+  appId: "1:662700376854:web:83529af948da8555dcc637",
+};
+
+export const app = initializeApp(firebaseConfig);
